feat: 스킨, 버튼, 물고기 거래, 퀘스트 관련 사운드 추가

+++ Fishing/Assets/Script/DataManager.cs
@@ -276,7 +276,7 @@ public class PlayerData {
 public class NpcQuest {
     public List<int> questList = new();
     public List<NormalQuest> normalQuests = new();
-    public float timer;
+    public int timer;
 }
 
 [Serializable]
